+++ src/app/(protected)/layout.tsx
