--- src/app/main/main.css
@@ -1,39 +1,0 @@
-/* .main{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    height: 100%;
-    overflow: hidden;
-}
-
-.main div{
-    flex:1;
-    align-items: center;
-    justify-items: center;
-}
-.store-img img{
-    height: 280px;
-}
-.main:last-child{
-    flex:none;
-}
-
-.good-bad{
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    max-width: 100%;
-    max-height: 40%;
-}
-.good-bad img{
-    margin: 10px 10px 10px 0px;
-    max-width: 25%;
-    max-height: 25%;
-}
-
-.under-menu-bar{
-    display: flex;
-}
-.under-menu-bar img{
-    max-width: 20%;
-} */
